разбила код в main.js на модули constants.js, helpers.js, create-photo.js
