+++ js/constants.js
@@ -1,0 +1,36 @@
+const NAMES = [
+  'Артем',
+  'Сан Саныч',
+  'Семён',
+  'Иван',
+  'Джон',
+  'Катя',
+  'Петр Петров',
+  'Добрыня',
+  'Алиса',
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const DESCRIPTION = [
+  'Не мог не выложить это фото...',
+  'Как вам обработка фотографии?',
+  'Люблю природу осенью',
+  'Моя обновка, зацените',
+  'Удачный кадр...',
+  'Наконец собрались с друзьями! Было весело!',
+  'Найдите меня на фотографии)))',
+];
+
+const MAX_COMMENTS_NUMBER = 4;
+
+const USER_PHOTOS_COUNT = 25;
+
+export {NAMES, MESSAGE, DESCRIPTION, MAX_COMMENTS_NUMBER, USER_PHOTOS_COUNT};
